[kmac] Fix silly lint warning in keccak_2share.sv
These parameters were added in 233ea81e5a9 and were supposed to be a
lint fix themselves. JasperGold complains about defining a "parameter"
when there is also a formal parameter declaration list and points out
that it will be magically turned into a localparam.

Do the transformation manually and silence the warning.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>