[hmac/prim_sha2,rtl] Implement SW error for invalid HMAC config
This adds Key_None to HMAC key_length to represent
all-zero or invalid key length configuration.
It also fixes the encoding for key_length
and digest_size to make them consistent, and updates
the documentation. This implements support for error signalling to SW
when an invalid digest size is configured, or an invalid key length (when HMAC
is enabled), or when a 1024-bit key length is configured
for SHA-2 256.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>