[ci] Move CW340 job to CW340 board in default pool
We can now select FPGAs by their demands from a single pool rather than
splitting them across two pools.

Needed because the CW340 in the default pool is in the hyperdebug config
while the one in `FPGA CW340` is still in the SAM3X config and we can't
remove it without breaking in-flight pull requests.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>