[reggen,fpv] Make pend_trans assumptions symmetrical
These assumptions weren't quite right before. They are trying to
ensure that we don't pass a silly index to the pend_trans array. But
we use both h2d.a_source and d2h.d_source as indices, so we need to
constrain both.

The "MAX_A_SOURCE" name is a bit dubious now, but I intend to
generalise it in the next commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>