[i2c] Add ACK Control Mode
Add an N-byte automatic ACK control so software may observe partial
command data and NACK the transaction in the middle of a transfer. The
counter begins at 0 for every transfer, and it decrements for every data
byte that is ACK'd. When the FSM reaches the (N)ACK phase, it stretches
the clock and awaits instructions from software (up to the stretch
timeout). Software can either program the Auto ACK Counter to a higher
number to ACK the pending byte (and subsequent ones, if the value
programmed is greater than 1), or it can explicitly NACK that byte and
the rest of the transaction.

Show the pending data byte in a separate CSR, away from the ACQ FIFO,
and change the "ACQ FIFO full" interrupt into a generic "stretching on
ACQuisition" interrupt, with the particular reason shown in the STATUS
CSR. When ACK Control Mode is disabled, that interrupt becomes purely an
"ACQ FIFO full" interrupt again.

A mechanism for mid-transfer NACKing is required for SMBus protocol
support.

Signed-off-by: Alexander Williams <awill@opentitan.org>