[reggen,fpv] Fix array sizes in fpv_csr.sv.tpl
Before and after this change, there are three arrays in this code that
are addressed by the index of an HRO register: regwen, access_policy
and exp_vals. The index of the HRO register comes from a variable
called hro_idx. Valid HRO registers get their position in the list of
such registers. Other values get a dummy value that is one greater.

Before this change, the exp_vals correctly had an extra element (to
represent this dummy hro_idx). The other two arrays did not.

This change makes things more uniform by adding the extra element to
those two arrays and initialising it to some default value. We should
no longer be able to access the arrays out of bounds.

The change also fixes up the width of hro_idx, avoiding some lint
errors.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>