[entropy_src,dv] Inline entropy_src_assert_if into vseq
This interface exposed a task called assert_off_err and that task was
only used in one place (in entropy_src_err_vseq). Unfortunately, the
interface wasn't technically conformant because it had an empty port
list and some tools throw out warnings about this.

Since we're just using the task in one place, it's rather simpler to
move the definition to the vseq and get rid of the interface entirely.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>