[reggen,fpv] Fix array sizes in fpv_csr.sv.tpl... again!
The previous change was mostly right, except for an off-by-one which
didn't get triggered with the HRO register count I was testing with.
If num_hro_regs = N, they will have indices 0, 1, ..., N-1. Then we'll
throw in an extra index for non-HRO registers, with value N.

To represent an arbitrary one of these indices, we need $clog2(N)
bits. My code was wrong and used $clog2(N+1) (because there are N+1
entries in the array).

It turns out that I was testing with a value of N where those two
calculations give the same result. But it doesn't work with N=3 (the
situation I have now).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>