[dvsim] Tweak mode find/merge to be simpler to understand
The existing code worked through a list of desired names and found
modes with these names. If the merge_modes flag was true then it also
merged each of the hits into the object supplied as the "mode"
argument. It turns out that this merging behaviour was actually only
used at one call-site.

This commit simplifies things. We now have a function called
find_mode_list which just does the filtering operation. The merging
part of things gets moved up to the call site.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>