[dvsim] Tweak find_and_merge_modes to be simpler to understand
The existing implementation worked through a list of desired names and
found modes with these names. If the merge_modes flag was true then it
also merged each of the hits into the object supplied as the "mode"
argument.

The change here should make things a bit simpler to understand. We now
filter the list first and then (if merge_modes is true) merge in each
hit afterwards.

Note that the "find_mode_list" helper function is now essentially a
filter (but taking the order from mode_names rather than modes).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>