[i2c,dv,cleanup] Refactor timing-related configuration into timing_pkg
This change creates a common home for all timing related routines, constraints
and constants. It contains a class 'i2c_timing_cfg' which can be instantiated
and randomized directly to create new configurations of timing parameters. This
class should be extended to add/disable/redefine constraints by child sequences
and other consumers.

Timing parameters are still randomized as part of the base vseq, but the whole
class instance is declared as rand.

Previously, the timing parameters (and their constraints) were defined in the
base sequence, and sub-sequences then override parts of these constraints, or
added new ones. This often led to code-duplication, and subtly different
constraints often with the same comment which caused confusion.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>