[sysrst_ctrl,sival] Fix ulp_z3_wakeup
The test was passing but this was due to luck: the test phase
was stored in RAM and used the OTTF framework for backdoor writes
but in fact the test goes to deep sleep so on wakeup the test phase
is wrong. This commits rework the code so that the device reads
the test phase over three wires that are set by the host.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 3d0303dc775f3b3d52771a97e4eb81955fab3e78)