[otlib,test_utils] Add a GPIO monitoring test util
This util makes it easier to create a monitoring object and manipulate
monitoring events. It also provides a way to dump a monitoring trace
as a VCD file in the log. This can be very useful during debugging and
also the monitor can print the VCD on drop so that if a test fails then
the trace is automatically printed which can be very useful when debugging
failures which are not easy to reproduce.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>