[i2c] Add StretchAcqSetup state to preserve ACK setup time
Add StretchAcqSetup state to follow the release of the hold from
StretchAcqFull after the ACQ FIFO gains the necessary space. Continue to
stretch SCL so the change to SDA doesn't violate setup time, then
proceed to the normal ACK setup state.

Remove the immediate write to the ACQ FIFO in the StretchAcqFull state
when we're proceeding to a normal ACK. This created extra ACQ FIFO
writes when it transitioned to the AcquireAckWait state, and the FSM
gets simpler and less error-prone if only error cases operate on FIFOs
outside of the non-stretch states.

Signed-off-by: Alexander Williams <awill@opentitan.org>