[prim,fpv] Only allow unconstrained counters in prim_count FPV
The counters in prim_count have a magic delta signal (fpv_force),
which can be used to make them jump around. This causes all manner of
problems in FPV tests for other blocks that happen to include
prim_count.

Tighten the ifndef line so that fpv_force is only allowed to be
nonzero when we're doing FPV for prim_count on its own. This way, the
outputs of the prim_count instances won't randomly jump around in
blocks that want to use them.

Honestly, I suspect that the fpv_force signal isn't sufficiently
constrained to be useful even in that situation. But that's probably
an issue for a different commit!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>