[sram_ctrl, rtl] Fix valid -> ready timing path
Previously the grant signal from the SRAM primitive was used directly to
calculate a tilelink ready signal. As the grant factors in this request
this produces a valid -> ready timing path. This causes timing issues
and is not required.

This commit refactors the logic to avoid using the SRAM grant to
calculate the tilelink ready.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>