[bazel] Enable support for ecdsa signatures.
Add support for ECDSA in Bazel infrastructure under signing.bzl. At the
moment, the signing infrastructure supports RSA and ECDSA. This is so
that we can migrate to ECDSA via incremental changes.

ECDSA signing can be tested as follows:

```shell
$ bazel build //signing/examples:digests_ecdsa
```

The `opentitantool image manifest verify` command can be used to verify
the signature over the pre-signed artifacts.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>