[opentitantool] Add ecdsa params to image commands.
Updates the image commands in opentitantool to support ECDSA public keys
and signatures. The current manifest implementation uses an union to
store RSA or ECDSA public keys and signatures. In the RSA case, the
public key/signature takes up the entire union space. In the ECDSA
case the key/signature occupy the first 256bits of the data
container and the rest is padded with a known pattern.

This is a temporary measure until we remove support for RSA.

The opentitantool manifest module was updated to use public key and
signature fields in favor over the RSA specific names.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>
(cherry picked from commit 4305e47ec8297aea66af2135dc79fa5788e55d0e)