[prim,fpv] Properly "waive" some unreachable prim_count assertions
This comes from an instance of prim_count which is created with
decr_en_i wired to zero. The prim_count module contains quite a few
assertions that say "If decr_en_i is one then ...", but these are all
unreachable!

Add a parameter to the prim module that waives the coverage, together
with an assertion that says this is really true.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>