[alert_handler,fpv] Move a consistency assertion from FPV to RTL
No change in behaviour here, but the previous code caused lint
warnings from Jasper. The problem is that we were trying to define a
constant (with the '{...} syntax), but the contents of the constant
contained "alert_handler_esc_timer.", so depended on the object we'd
built.

These hierarchical references were only needed because we were trying
to assert things about the module's internal state. The assertion
makes perfect sense in the context of the module itself, so let's just
move it there.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>