[usbdev] Refinement to sampling logic
Avoid sampling when a transition has just been detected.

When the usbdev clock frequency is higher than that of the host it
has a tendency to 'get ahead', and when the differential signals are
captured in CDC on different sides of the clock edge (induced in
simulation by prim_cdc_rand_delay) a transition may be observed
when moving from phase 1 to phase 2, which caused 00 to be sampled
into the line_history, followed by a further valid sample just two
cycles later.

Observed with usbdpi running at 47982kHz cf usbdev 48MHz.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>