[prim,fpv] Rephrase some "backwards" assertions in prim_count
In this module, we have forwards assertions that say "If an input
happens then we change accordingly" and we have backwards assertions
that say "If we change then an input must have been asserted".

We've got several different inputs that might change the
count (incr/decr/set/clear). The existing backwards assertions were
quite strong and asked for a change that definitely wasn't caused by a
different input. For example, ClrBkwd_A said "If we change and it
wasn't caused by incr, decr or set, then we must have seen clear".

This is fine, but it creates a cover property to see a change that
isn't caused by the other inputs. We might not be able to hit
that! In particular, in the alert_handler_esc_timer module, we weren't
every seeing set_i without incr.

But we don't really care! This commit generalises the backward
assertion. This should give the same level of confidence, but avoids
the vacuity that Jasper reports as a coverage hole otherwise.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>