[alert_handler,rtl] Correct width of prim_count instance
This width wasn't quite right for u_prim_count_esc_cnt in the
design (and Jasper correctly points out that we can't get to the
maximal count value).

We pick a target to ping by incrementing that target's index every
time we went around the FSM that drove the pings. Of course, there are
only finitely many targets (N_ESC_SEV of them) and we wrap at
N_ESC_SEV-1.

With the current code (before or after this commit), this is done by
driving an esc_cnt_clr signal. Before we switched to use
prim_count (in commit b5eb150), the wrapping was implemented directly
in this file.

The switch to prim_count was reasonable, but instantiated the counter
with a width of PING_CNT_DW, which is 16, instead of the number of
bits needed to represent N_ESC_SEV-1 = 3, which is just 2.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>