[flash_ctrl] Pop read pipeline FIFOs for data and mask in sync
This commit fixes a bug in the read pipeline where the data and the mask
storage FIFO could get out of sync in case of the mask computation
running behind the actual Flash read (e.g. due to back pressure in the
shared scrambling logic).

To avoid the FIFOs getting out of sync, the design now tracks whether
for reads that don't require descrambling, a mask was still computed.
This can happen if for scramble enabled locations that are being erased.
In case a mask computation has been started, the FIFOs are now only
popped if both the data and the mask FIFO contain a valid entry (i.e.
the mask computation for the data to be popped has finished).

This resolves lowRISC/OpenTitan#22443

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>