[rom_ctrl,lc_ctrl,dv] Specialise scratch_path/rel_path for variants
Without this change, there is a problem where we have two different
variants of each block that are both getting built for nightly tests
and both variants write results to the same place. This fix nicks the
equivalent code from the aes block (which has to solve the same
problem).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>