[edn,rtl] Fix assertions
The existing code wasn't quite right and actually causes a compilation
error when you try to push it through Jasper. For a bit more detail,
edn_enable_fo is the "edn enable fanout". This is indexed by entries
from FatalErr(=1) to LastEdnEntry-1, and doesn't have anything at
index zero. (This is what causes the compilation error)

In addition, one of the assertions no longer got disabled upon disabling
EDN. As a result, it would always fire after re-enabling EDN which
intentionally clears the FIFOs the assertion is checking.

These bugs got introduced by a974108, which removed an "enable_q" signal
from edn.sv. Looking at the assertions, I think the property that
we're talking about is "some of EDN is enabled". To get what I hope is
the intended behaviour, I've used a unary OR to combine together all
of the enable signals.

Co-authored-by: Pirmin Vogel <vogelpi@lowrisc.org>
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>