[tlul,rtl] Rework tlul_err_resp
Previously the logic of tlul_err_resp was more complex than was required
(having seperate flops for tracking a pending request and response where
it's sufficient to just track a pending response).

This simplifies the logic and cuts a path from the incoming data ready
to the outgoing address ready.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>