[rv_dm,dv] Slightly tidy up interface connections in tb.sv
No functional change, except that we drop an unnecessary trailing '*'
sign in a couple of the paths. But it took me a few minutes to work
out what was using each of these connections! So I've left a couple of
comments that should make it a bit more obvious.

I've also re-ordered the lines to group related interfaces a bit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>