[rom] Avoid touching watchdog config on wake from sleep
When waking from sleep, the watchdog IP will not have been reset, so it
is not necessary to re-configure it.  In fact, if watchdog REGWEN has
been used to disable writes to its registers, attempting to do so can
generate an exception causing chip reset, which will reset not only the
watchdog, but other IP blocks that the owner code wanted retained across
the period of sleep.

Also, do not re-configure pinmux or attempt to read SW_BOOTSTRAP pins if
waking from sleep.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I503fd65f2203c425ae01ea7044479533aeac8895