[rv_dm,rtl] Remove next_dm_addr_i port
This port was added to the RISC-V debug module that we vendor from
pulp in order to support having a chain of debug modules. I don't
believe we're planning to build anything like this in OpenTitan in the
near term, so let's just tie it to zero rather than exposing the port
from the block.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>