[i2c] Add bus arbitration support for the controller
Detect when the controller is transmitting SDA high, but the value on
the wire is low, resulting in lost arbitration in a multi-controller
environment.

Proceed immediately to Idle, set a bit in CONTROLLER_EVENTS for lost
arbitration, and halt at Idle until that bit is cleared. This allows
software to respond and prepare the FMT FIFO for the next transmission,
which may be a retry.

Add a parameter for the top-level input delay between cio_sda_o and
cio_sda_i, and avoid triggering bus errors during that period.
Internally, the detection window will also be reduced by the rise time
for SCL, so this parameter only represents the worst case delay from the
output flop's output to the pad plus the pad back to the first input
flop's input.

Set InputDelayCycles to 1 for earlgrey. Assume a 50 ns glitch filter
contributing to input delay and that the delay above the 24 MHz clock's
period isn't enough to push the sampling point into the next clock cycle
(i.e. past the second edge after the output edge).

Signed-off-by: Alexander Williams <awill@opentitan.org>