[prim,fpv] Rephrase prim_count error assertions
The correctness assertion is unchanged (because (A |-> B) && (B |-> A)
is true if (A == B) when A and B are expressions). The change here is
in the generated cover properties.

The correctness property means that we are proving there's only one
cover point that is needed (let's make it "err_o" for brevity). But we
will only see this happen when the counters get forced. Fortunately,
we have a way of doing that (see the fpv_force variable). Only
generate the cover property in that situation.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>