[chip tests] Fix kmac_app_rom_test test
Use the (test)ROM hash files to compare the digest computed by
ROM_CTRL with the one computed by the scrambler. This requires
a bit of boilerplate because there is a different hash file for
different execution environment so we cannot use a single
`opentitan_test` rule to make that work. Instead create one
test per execution environment and collect them in a test suite
manually. This also explains a slightly unusual path for the
`sw_images` in the DV test.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>