[prim,fpv] Correct assertions for commit_i input
The ClrFwd_A assertion says "When a clear is requested, it should take
effect". But this isn't quite true because commit_i might be false,
which would mean the clear operation wouldn't take effect in cnt_q.
Strengthen the assumption to "When a clear is requested and we really
mean it ..."

Similarly for SetFwd_A.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>