[pentest] Constant trigger window for OTBN
For side-channel and fault injection analysis, a constant trigger
window is required. However, currently, for OTBN analysis, this is
not the case as entropy complex reseeds are issued during OTBN execution.

This commit adds the SCA & FI specific sca_configure_entropy_source_otbn()
function. This function is a copy of entropy_testutils_auto_mode_init()
but sets the reseed_interval of EDN0...4 to the max. available value,
i.e., 0xffffffff.

With this change, the trigger window stays constant for OTBN measurements.

Thanks @jadephilipoom for helping to debug this!

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>