[i2c] Prevent transmission until the bus is free
Consolidate start/stop detection, bus idle timeouts, and transaction
boundary checking in a separate bus monitor module. Also move counting
of the bus free time to the bus monitor, and use all of these inputs to
determine when a bus is free, ready for the controller's next
transmission.

Adjust the controller FSM to respond to the bus monitor's "bus free"
status signal.

Signed-off-by: Alexander Williams <awill@opentitan.org>