[i2c] Prevent transmission until the bus is free
Consolidate start/stop detection, bus idle timeouts, and transaction
boundary checking in a separate bus monitor module. Also move counting
of the bus free time to the bus monitor, and use all of these inputs to
determine when a bus is free, ready for the controller's next
transmission.

Adjust the controller FSM to respond to the bus monitor's "bus free"
status signal.

Remove the "SDA is low in Idle" check from the SDA interference
interrupt. There are various single-controller scenarios where this may
trigger uselessly, and it isn't compatible with multi-controller
scenarios. If a user wishes to know if SDA is low when the RTL's comment
literally indicated, merely around the time when the controller is first
enabled, then software may probe the RX oversample CSR for the current
value.

Signed-off-by: Alexander Williams <awill@opentitan.org>