[dv] spi_device_pass_base_vseq backpressure fix
If there's back pressure in the D-channel after accepting the written
data, a call to 'csr_rd' will block until the D-channel response is
output.
This causes issues when checking if the BUSY bit is unset, since it
can happen the flash_status write cleared the BUSY bit, but since the
D-channel response took long to be output by that time the BUSY bit
may have been set again.

So, if the BUSY bit is unset early due to backpressure, we don't check
that it's still set at a later point, since that can mean the BUSY bit
has been set again after being unset.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>