[pinmux] Copy "countermeasures" list to pinmux.hjson
This is a bit of a mess because the block is top-specific but isn't
quite converted to be generated with ipgen yet. The list of
countermeasures was already added to pinmux.hjson.tpl (the source for
the top-specific hjson files), but we've got some FPV testing which is
based on the original pinmux.hjson.

A proper fix would be to get rid of this file entirely and generate
pinmux with ipgen. But that's a reasonable chunk of work. This change
is trivial and gets FPV tests working again...

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>