[manuf] increase reliability of lc read testutil
The LC read test utility was not waiting for the lc_ctrl to be ready to
receive transactions before reading the LC state. This increases the
reliability of this utility, and test cases that use it, specifically,
the `ft_provision` test case.

Additionally, this move the `ft_provision` silicon exec_env to the
`teacup` interface (a minor cleanup).

Signed-off-by: Tim Trippel <ttrippel@google.com>