[i2c] Halt subsequent read if prior read failed
Add a CSR to report fault events that occurred during a read, and gate
the next read (stretch clock at the first bit) until the event latch is
cleared. This is for reads to this IP's target module.

The gate helps prevent stale data from emerging from the FIFO.

Also roll in some of the functionality the TXRST_ON_COND CSR was meant
to provide, and make this a catch-all gate for releasing the TX FIFO on
a read instead. The prior mechanism had some corner cases that produced
races with software, such as software being able to inadvertently write
a new (yet stale) byte of data after the clear yet before being notified
that the command completed. The separate gating CSR helps avoid such
races by not allowing software's normal write path to bypass event
handling.

Signed-off-by: Alexander Williams <awill@opentitan.org>