[ot_certs] Rework Name handling
Names were previously represented as maps but this is problematic
for two reasons:
- order is not preserved,
- it cannot represente multi-valued RDNs (although they are are).

This commit changes the representation to an array of maps so that
it maps cleanly to the ASN1 representation (sequence of sets). In
particular the order in the array preserved while the order in the
map/set is arbitrary (this matches the expected X501 behaviour).
This makes the hjson a little bit more ugly but it's not too terrible.
One issue that arises from this change is that although it is very
easy to modify the builder to handle multi-valued RDNs, the openssl
parser cannot reconstruct them because the openssl-sys binding
misses a function (see FIXME in the code). Since multi-valued RDNs
are rare and not used in our codebase presently, we simply assume
single-valued RDNs are the moment.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 11e5cf55b7871c1757376b784fc4538fc5764563)