[opentitantool] Re-attach kernel driver
Sometimes, tools using libusb may ask that the kernel driver for a
particular USB interface be "detached" from that interface, such that
libusb can be used to control which data transfers to make.  At Google,
"servod" is one such tool, which has a habit of detaching the kernel
serial port driver from C2D2 or uServo debuggers, meaning that the
/dev/ttyUSBn devices would not be present.

This CL adds logic to make sure that the kernel serial driver is
attached to any HyperDebug USB interface, before searching for the
proper devUSBn.

Change-Id: I900532924df8f46467e904f68ff411349f72e3fe
Signed-off-by: Jes B. Klinke <jbk@chromium.org>