[sival, rv_core_ibex] fix ibex address translation test
Instead of map arbitrary address spaces (which requires setting up PMP),
map within .text sections. The illegal access fault is replaced with
illegal instruction fault when the address is not mapped.

This allows the test to run inside sival env (but not sival_rom_ext, since
address translation is disabled).

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>