[sram_ctrl] Mubi encoding for write enable
This commit hardens the write enable signal used in the SRAM logic
against fault attacks by encoding it with a mubi.

The encoded WE signal is generated in the tlul_adapter_sram, passed
through sram_ctrl, and forwarded to the prim_ram_1p_scr module.
In this module, a integrity check is conducted, triggering an error
on failure.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>